+++ docs-site/.vitepress/theme/console/pages/AITools.vue
@@ -1,4 +1,0 @@
-<template>
-  <div>AITools Page</div>
-</template>
-