+++ docs-site/.vitepress/theme/console/pages/Logs.vue
@@ -1,4 +1,0 @@
-<template>
-  <div>Logs Page</div>
-</template>
-