+++ docs-site/.vitepress/theme/console/pages/Memories.vue
@@ -1,4 +1,0 @@
-<template>
-  <div>Memories Page</div>
-</template>
-