+++ docs-site/.vitepress/theme/console/pages/Modules.vue
@@ -1,4 +1,0 @@
-<template>
-  <div>Modules Page</div>
-</template>
-