+++ docs-site/.vitepress/theme/console/pages/Status.vue
@@ -1,4 +1,0 @@
-<template>
-  <div>Status Page</div>
-</template>
-