+++ docs-site/.vitepress/theme/console/pages/Teams.vue
@@ -1,4 +1,0 @@
-<template>
-  <div>Teams Page</div>
-</template>
-